Updated Email System
Now, if a user (patient or doctor) forgets their password, they can go to the reset-password.html page to reset it. If they choose to reset it, they will receive an email containing the reset link.
If they reach the page without clicking this link, the password change will not work.