Secure chat encryption
A feature has been added that allows encrypted chat messages to be saved in the database. At the same time, thanks to the REST API, it is possible to display the decrypted chat content on the frontend. Currently, the functionality that returns the decrypted chat to the frontend is not working